requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.6
pip==21.3.1
requests==2.30.0
setuptools==60.2.0
tqdm==4.65.0
urllib3==2.0.2
wheel==0.37.1
yarl==1.9.2
python-dotenv==0.21.1
